feat(88x2bu): add driver RTL8812BU/RTL8822BU v5.13.1
Although there is a driver for RTL8822BU in newer kernels (>=6.2),
this kernel is not available with current Yocto releases. This kernel
version has not been released yet, either. Hence, it is useful to
have a driver for this REALTEK chipset at hand for current kernels.

The Linux driver pages does not state whether RTL8812BU is supported
or not. So, maybe this driver is good to have for these chipsets too.

see: https://wireless.wiki.kernel.org/en/users/drivers/rtl819x